+++ pareg_core/src/proc/mod.rs
@@ -1,6 +1,7 @@
 //! This module contains raw implementation of proc macros with `proc_macro2`.
 
+mod err;
 mod from_arg;
 mod parsef;
 
-pub use self::{from_arg::*, parsef::*};
+pub use self::{err::*, from_arg::*, parsef::*};